backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
filetype==1.2.0
Pillow>=10.4.0
# psycopg2-binary>=2.9.10   # Uncomment for PostgreSQL (prod). SQLite used in dev.

